+++ src/modules/users/dtos/create-user.dto.ts
@@ -3,12 +3,12 @@ import { IsString, IsNotEmpty } from 'class-validator';
 export class CreateUserDto {
   @IsString()
   @IsNotEmpty()
+  email: string;
+
+  @IsString()
+  @IsNotEmpty()
   nickname: string;
 
   @IsString()
   introduce: string;
-
-  @IsString()
-  @IsNotEmpty()
-  code: string;
 }
